+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## [Unreleased]
+
+### Changed
+
+- concord-server: fixed a bug causing incorrect matching of
+Concord repositories with incoming GitHub events.
+
+
+
 ## [0.96.0] - 2018-11-07
 
 ### Added